Satisfy error log structural checks
